refactored folder scheme for the component create task to be part of the home page
