+++ src/app/user-logged/components/navbar/navbar.component.html
@@ -9,12 +9,6 @@
             <li class="nav-item active">
                 <a class="nav-link" href="#">Home</a>
             </li>
-            <li class="nav-item">
-                <a class="nav-link" href="#">Games</a>
-            </li>
-            <li class="nav-item">
-                <a class="nav-link" href="#">Series/Animes</a>
-            </li>
         </ul>
         <span class="navbar-text mr-4">
             <a href="">Settings</a>
