+++ src/app/user-logged/user-logged.component.html
@@ -3,9 +3,7 @@
 <div class="container-fluid">
   <div class="row">
     <div class="col"></div>
-    <div class="col text-center">
-      <app-create-task-form></app-create-task-form>
-    </div>
+    <div class="col"></div>
     <div class="col"></div>
   </div>
     <router-outlet></router-outlet>
