+++ src/app/user-logged/user-logged.module.ts
@@ -5,7 +5,7 @@ import { UserLoggedRoutingModule } from './user-logged-routing.module';
 import { UserLoggedComponent } from './user-logged.component';
 import { HomeComponent } from './pages/home/home.component';
 import { TaskComponent } from './pages/home/components/task/task.component';
-import { CreateTaskFormComponent } from './components/create-task-form/create-task-form.component';
+import { CreateTaskFormComponent } from './pages/home/components/create-task-form/create-task-form.component';
 import { NavbarComponent } from './components/navbar/navbar.component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
